+++ CHANGES.rst
@@ -36,6 +36,9 @@ Changes:
 - Avoid ``Job.progress`` updates following ``failed`` or ``dismissed`` statuses to keep track of the last real progress
   percentage that was reached when that status was set.
 - Improve typing of database and store getter functions to infer correct types and facilitate code auto-complete.
+- Implement ``Job`` `dismiss operation <https://docs.ogc.org/DRAFTS/18-062.html#sec_cons_dismiss>`_ ensuring
+  pending or running tasks are removed and output results are removed from disk.
+- Implement expected HTTP Gone (410) status from already dismissed ``Job`` when requested again.
 
 Fixes:
 ------